+++ src/fixtures/billingEVent/billingEvent.test.false.js
@@ -1,0 +1,15 @@
+module.exports = {
+  validate: {
+    params: {
+      optimizationGoal: 'APP_INSTALLS',
+      billingEvent: 'OFFER_CLAIMS',
+    },
+    response: false,
+  },
+  map: {
+    params: {
+      optimizationGoal: 'APP_INSTALLS',
+    },
+    response: ["IMPRESSIONS"]
+  }
+}