+++ src/fixtures/billingEVent/billingEvent.test.true.js
@@ -1,0 +1,15 @@
+module.exports = {
+  validate: {
+    params: {
+      optimizationGoal: 'APP_INSTALLS',
+      billingEvent: 'IMPRESSIONS',
+    },
+    response: true,
+  },
+  map: {
+    params: {
+      optimizationGoal: 'APP_INSTALLS',
+    },
+    response: ["IMPRESSIONS", "APP_INSTALLS"]
+  }
+}