+++ src/fixtures/optimizationGoal/optimizationGoal.test.false.js
@@ -1,0 +1,15 @@
+module.exports = {
+  validate: {
+    params: {
+      campaignObjective: 'BRAND_AWARENESS',
+      optimizationGoal: 'POST_ENGAGEMENT',
+    },
+    response: false,
+  },
+  map: {
+    params: {
+      campaignObjective: 'BRAND_AWARENESS',
+    },
+    response: ["BRAND_AWARENESS"]
+  }
+}