+++ src/fixtures/optimizationGoal/optimizationGoal.test.true.js
@@ -1,0 +1,15 @@
+module.exports = {
+  validate: {
+    params: {
+      campaignObjective: 'BRAND_AWARENESS',
+      optimizationGoal: 'BRAND_AWARENESS',
+    },
+    response: true,
+  },
+  map: {
+    params: {
+      campaignObjective: 'BRAND_AWARENESS',
+    },
+    response: ["BRAND_AWARENESS", "REACH"]
+  }
+}